requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sqlalchemy>=2.0.23
pydantic>=2.5.1
pydantic-settings>=2.1.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
python-multipart>=0.0.6
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.1
pytest-mock>=3.12.0
alembic>=1.13.1